Add nick user events module
